requirements.txt
feedparser==6.0.11
google-genai==1.10.0
python-telegram-bot==21.10
python-dotenv==1.1.0
